Create Calculator
The start of my Calculator project. First, it prompts the user to enter two numbers using the 'nextDouble' method of the Scanner class. Then, it displays a list of available operations and prompts the user to select one using the next() method. It uses a switch statement to determine which operation to perform, and we handle division by zero as a special case. Finally, it displays the result using the println() method. When you run this program, it will prompt the user to enter two numbers and select an operation, perform the operation on the numbers, and display the result.